[Feat] #42 - View 전환 관련 코드 작성
