+++ Treehouse/Treehouse/Presentation/Auth/RegisterScene/Views/SetUserIdView.swift
@@ -15,53 +15,78 @@ struct SetUserIdView: View {
     
     // MARK: - State Property
     
+    @Environment(ViewRouter.self) var viewRouter: ViewRouter
+    @State var viewModel = UserSettingViewModel()
+    
     @State var userId: String = ""
-    @State var isButtonEnabled: Bool = false
     @State var textFieldState: TextFieldStateType = .notFocused
     @FocusState private var focusedField: SetUserIdField?
     
     // MARK: - View
     
     var body: some View {
-        VStack(spacing: 0) {
-            VStack(alignment: .leading, spacing: 24) {
-                Text(StringLiterals.Register.registerTitle4)
-                    .fontWithLineHeight(fontLevel: .heading1)
-                    .foregroundStyle(.black)
+        @Bindable var viewRouter = viewRouter
+        
+        NavigationStack(path: $viewRouter.path) {
+            VStack(spacing: 0) {
+                VStack(alignment: .leading, spacing: 24) {
+                    Text(StringLiterals.Register.registerTitle4)
+                        .fontWithLineHeight(fontLevel: .heading1)
+                        .foregroundStyle(.black)
+                    
+                    Text(StringLiterals.Register.guidanceTitle3)
+                        .fontWithLineHeight(fontLevel: .body3)
+                        .foregroundStyle(.gray5)
+                }
+                .frame(maxWidth: .infinity, alignment: .leading)
+                .padding(.trailing, 2)
+                .padding(.bottom, 26)
                 
-                Text(StringLiterals.Register.guidanceTitle3)
-                    .fontWithLineHeight(fontLevel: .body3)
-                    .foregroundStyle(.gray5)
+                userNameTextField
+                
+                Spacer()
+                
+                Button(action: {
+                    if viewModel.isButtonEnabled {
+                        viewModel.userId = userId
+                        viewRouter.push(RegisterRouter.showUserProfileView)
+                    }
+                }) {
+                    Text(StringLiterals.Register.buttonTitle5)
+                        .padding()
+                        .frame(maxWidth: .infinity)
+                        .frame(height: 56)
+                        .foregroundStyle(viewModel.isButtonEnabled ? .gray1 : .gray6)
+                        .background(viewModel.isButtonEnabled ? .treeBlack : .gray2)
+                        .cornerRadius(10)
+                        .padding(.trailing, 1)
+                }
             }
-            .padding(.leading, 2)
-            .padding(.bottom, 26)
-            
-            userNameTextField
-            
-            Spacer()
-            
-            Button(action: {
-                
-            }) {
-                Text(StringLiterals.Register.buttonTitle5)
-                    .padding()
-                    .frame(maxWidth: .infinity)
-                    .frame(height: 56)
-                    .foregroundStyle(isButtonEnabled ? .gray1 : .gray6)
-                    .background(isButtonEnabled ? .treeBlack : .gray2)
-                    .cornerRadius(10)
+            .padding(EdgeInsets(top: 22, leading: 24, bottom: 30, trailing: 24))
+            .toolbar {
+                ToolbarItem(placement: .topBarLeading) {
+                    Button(action: {
+                        
+                    }) {
+                        Image(systemName: "chevron.left")
+                            .foregroundColor(.treeBlack)
+                    }
+                    .padding(.top, 5)
+                }
+            }
+            .navigationDestination(for: RegisterRouter.self) { router in
+                viewRouter.buildScene(inputRouter: router, viewModel: viewModel)
             }
         }
-        .padding(EdgeInsets(top: 66, leading: 24, bottom: 23, trailing: 22))
         .onAppear {
             UITextField.appearance().clearButtonMode = .whileEditing
         }
         .onChange(of: userId) { _, newValue in
             if self.isValidInputUserId(newValue) {
-                isButtonEnabled = true
+                viewModel.isButtonEnabled = true
                 textFieldState = .enable
             } else {
-                isButtonEnabled = false
+                viewModel.isButtonEnabled = false
                 textFieldState = .unable
             }
         }
@@ -107,5 +132,8 @@ private extension SetUserIdView {
 // MARK: - Preview
 
 #Preview {
-    SetUserIdView()
+    NavigationStack {
+        SetUserIdView(viewModel: UserSettingViewModel())
+            .environment(ViewRouter())
+    }
 }
